rfengine: offload raft entries to limit memory usage (part 2) (#2834)
This is part 2 of the raft entry offloading implementation (part 1:
#2812). This PR introduces two key fields to rlog files: `epoch_id` and
`last_term`, both stored in the manifest metadata. These fields are 
essential to the offloading mechanism:

Each rlog file and its entries are now tagged with an `epoch_id`, which
corresponds to the epoch of the WAL file from which the rlog file was
generated. This allows us to define an `offload_epoch` and then offload
any raft entries with `epoch_id <= offload_epoch`. The `offload_epoch`
will be derived from the current epoch, based on a configurable memory
limit introduced in a later PR.

Before offloading, a critical safety check must be performed: the term
of the rlog file's last entry must match the in-memory entry at the
same index. This ensures we don’t offload entries that have been 
overridden in memory by newer entries with higher terms, which could 
lead to data loss.

Signed-off-by: Bisheng Huang <hbisheng@gmail.com>