*: cherry pick tikv 15747, 15749, 15658, 16621 related to txn status (#2755)
* tikv_util: Support customizing evict policy and operations without promoting for LruCache (#15747)

ref tikv/tikv#11187

This PR makes the `LruCache` in `tikv_util` to support customizing how to determine an entries in the cache should be evicted.
This is part of solving the issue #11187, which needs a `TxnStatusCache`. The `TxnStatusCache` is desinged to use `LruCache` internally, with ability to get or insert items without promoting items to the head (most-recently-used) position. This PR adds `get_no_promote` and `insert_if_not_exist` functions to `LruCache`.

Signed-off-by: MyonKeminta <9948422+MyonKeminta@users.noreply.github.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>
Signed-off-by: ekexium <eke@fastmail.com>

* storage/txn: Add txn_status_cache module (#15749)

ref tikv/tikv#11187

This PR adds a module named txn_status_cache. It provides the type `TxnStatusCache`, which will be then put into scheduler to cache start_ts->commit_ts information of recently-committed transactions. This is part of fixing issue #11187 and will also be used for possible future optimizations such as checking txn status locally.

Signed-off-by: MyonKeminta <9948422+MyonKeminta@users.noreply.github.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>
Signed-off-by: ekexium <eke@fastmail.com>

* txn: Fix to the prewrite requests retry problem by using TxnStatusCache (#15658)

ref tikv/tikv#11187

Signed-off-by: MyonKeminta <9948422+MyonKeminta@users.noreply.github.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>
Signed-off-by: ekexium <eke@fastmail.com>

* txn: Fix the issue that CheckTxnStatus didn't make rollback on optimistic transaction's primary protected, which may break transaction atomicity (#16621)

close tikv/tikv#16620

Fix the issue that CheckTxnStatus didn't make rollback on optimistic transaction's primary protected, which may break transaction atomicity

Signed-off-by: MyonKeminta <9948422+MyonKeminta@users.noreply.github.com>

Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>
Signed-off-by: ekexium <eke@fastmail.com>

---------

Signed-off-by: ekexium <eke@fastmail.com>
Co-authored-by: MyonKeminta <9948422+MyonKeminta@users.noreply.github.com>
Co-authored-by: ti-chi-bot[bot] <108142056+ti-chi-bot[bot]@users.noreply.github.com>