rfengine: fix wal double write data loss (#3474)
When the wal is not rotated for one day, it will be force rotated, and the two wal writer may not rotate at the same time.
The primary wal rotated first, and the secondary wal rotated at the next write batch, then we mistakenly rotated twice.
Then compact wal just iterated 8K of data. cause data loss.

To make the rotation at the same time, we remote the force rotate logic.

Signed-off-by: Evan Zhou <coocood@gmail.com>