[Pick] rfengine: support double wal writer #3248 (#3705)
* rfengine: support double wal writer (#3248)

Signed-off-by: Evan Zhou <coocood@gmail.com>
Co-authored-by: Ping Yu <yuping@pingcap.com>

* resolve conflicts

Signed-off-by: 3pointer <luancheng@pingcap.com>

* fix case

Signed-off-by: 3pointer <luancheng@pingcap.com>

* rfengine: fix wal double write data loss (#3474)

When the wal is not rotated for one day, it will be force rotated, and the two wal writer may not rotate at the same time.
The primary wal rotated first, and the secondary wal rotated at the next write batch, then we mistakenly rotated twice.
Then compact wal just iterated 8K of data. cause data loss.

To make the rotation at the same time, we remote the force rotate logic.

Signed-off-by: Evan Zhou <coocood@gmail.com>

* fix test

Signed-off-by: 3pointer <luancheng@pingcap.com>

* fix case: test_native_lightweight_backup

Signed-off-by: 3pointer <luancheng@pingcap.com>

* fix format

Signed-off-by: 3pointer <luancheng@pingcap.com>

---------

Signed-off-by: Evan Zhou <coocood@gmail.com>
Signed-off-by: 3pointer <luancheng@pingcap.com>
Co-authored-by: Evan Zhou <coocood@gmail.com>
Co-authored-by: Ping Yu <yuping@pingcap.com>
Co-authored-by: Neil Shen <overvenus@gmail.com>